requirements.txt
django==4.1.1
djangorestframework==3.13.1
chatterbot_corpus==1.2.0

